# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zenith Staff Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
  message?: string;
}

export interface DiscordUser {
  id: string;
  username: string;
  discriminator: string;
  avatar?: string | null;
  avatarUrl: string;
}

export interface DiscordGuild {
  id: string;
  name: string;
  icon?: string | null;
  iconUrl?: string | null;
  isSetup: boolean;
}

export interface Guild {
  id: string;
  name: string;
  iconUrl?: string | null;
  isPremium: boolean;
  staffRoleId?: string | null;
  managementRoleId?: string | null;
  logChannelId?: string | null;
  applicationChannelId?: string | null;
  applicationReviewChannelId?: string | null;
  welcomeChannelId?: string | null;
  embedColor: string;
  embedFooter?: string | null;
  customBotName?: string | null;
  createdAt: string;
}

export interface UpdateGuildBody {
  staffRoleId?: string | null;
  managementRoleId?: string | null;
  logChannelId?: string | null;
  applicationChannelId?: string | null;
  applicationReviewChannelId?: string | null;
  welcomeChannelId?: string | null;
  embedColor?: string;
  embedFooter?: string | null;
  customBotName?: string | null;
  customBotAvatar?: string | null;
}

export interface ActivityItem {
  type: string;
  description: string;
  timestamp: string;
  actorUsername: string;
}

export interface GuildStats {
  totalStaff: number;
  activeStaff: number;
  pendingApplications: number;
  activeStrikes: number;
  activeLoas: number;
  totalPromotions: number;
  recentActivity: ActivityItem[];
}

export interface StaffMember {
  id: string;
  guildId: string;
  discordId: string;
  discordUsername: string;
  discordAvatarUrl?: string | null;
  robloxUsername?: string | null;
  rankId?: string | null;
  rankName?: string | null;
  divisionId?: string | null;
  divisionName?: string | null;
  callsign?: string | null;
  isActive: boolean;
  strikeCount: number;
  joinedAt: string;
  notes?: string | null;
}

export interface StaffListResponse {
  staff: StaffMember[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateStaffBody {
  discordId: string;
  discordUsername: string;
  discordAvatarUrl?: string | null;
  robloxUsername?: string | null;
  rankId?: string | null;
  divisionId?: string | null;
  callsign?: string | null;
}

export interface UpdateStaffBody {
  rankId?: string | null;
  divisionId?: string | null;
  callsign?: string | null;
  robloxUsername?: string | null;
  isActive?: boolean;
  notes?: string | null;
}

export interface Rank {
  id: string;
  guildId: string;
  name: string;
  discordRoleId?: string | null;
  position: number;
  color?: string | null;
  isDefault: boolean;
  createdAt: string;
}

export interface CreateRankBody {
  name: string;
  discordRoleId?: string | null;
  position?: number;
  color?: string | null;
  isDefault?: boolean;
}

export interface UpdateRankBody {
  name?: string;
  discordRoleId?: string | null;
  position?: number;
  color?: string | null;
  isDefault?: boolean;
}

export interface Division {
  id: string;
  guildId: string;
  name: string;
  description?: string | null;
  discordRoleId?: string | null;
  channelId?: string | null;
  color?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateDivisionBody {
  name: string;
  description?: string | null;
  discordRoleId?: string | null;
  channelId?: string | null;
  color?: string | null;
}

export interface UpdateDivisionBody {
  name?: string;
  description?: string | null;
  discordRoleId?: string | null;
  channelId?: string | null;
  color?: string | null;
  isActive?: boolean;
}

export interface ApplicationAnswer {
  question: string;
  answer: string;
}

export interface Application {
  id: string;
  guildId: string;
  applicantDiscordId: string;
  applicantUsername: string;
  robloxUsername?: string | null;
  status: string;
  answers: ApplicationAnswer[];
  reviewerId?: string | null;
  reviewerUsername?: string | null;
  reviewNotes?: string | null;
  submittedAt: string;
  reviewedAt?: string | null;
}

export interface ApplicationListResponse {
  applications: Application[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ReviewApplicationBody {
  notes?: string | null;
}

export interface ApplicationQuestion {
  id: string;
  guildId: string;
  question: string;
  placeholder?: string | null;
  isRequired: boolean;
  position: number;
  createdAt: string;
}

export interface CreateApplicationQuestionBody {
  question: string;
  placeholder?: string | null;
  isRequired?: boolean;
  position?: number;
}

export interface UpdateApplicationQuestionBody {
  question?: string;
  placeholder?: string | null;
  isRequired?: boolean;
  position?: number;
}

export interface Strike {
  id: string;
  guildId: string;
  staffId: string;
  issuedById: string;
  issuedByUsername: string;
  severity: string;
  reason: string;
  evidence?: string | null;
  isActive: boolean;
  appealStatus: string;
  issuedAt: string;
  expiresAt?: string | null;
}

export type CreateStrikeBodySeverity =
  (typeof CreateStrikeBodySeverity)[keyof typeof CreateStrikeBodySeverity];

export const CreateStrikeBodySeverity = {
  warning: "warning",
  strike: "strike",
  final_warning: "final_warning",
} as const;

export interface CreateStrikeBody {
  staffId: string;
  issuedById: string;
  issuedByUsername: string;
  severity?: CreateStrikeBodySeverity;
  reason: string;
  evidence?: string | null;
}

export interface RemoveStrikeBody {
  removedById: string;
}

export interface Loa {
  id: string;
  guildId: string;
  staffId: string;
  reason: string;
  status: string;
  reviewerId?: string | null;
  reviewerUsername?: string | null;
  reviewNotes?: string | null;
  startDate?: string | null;
  endDate?: string | null;
  requestedAt: string;
  reviewedAt?: string | null;
}

export interface CreateLoaBody {
  staffId: string;
  reason: string;
  startDate?: string | null;
  endDate?: string | null;
}

export interface ReviewLoaBody {
  reviewerId: string;
  reviewerUsername: string;
  notes?: string | null;
}

export type PromotionType = (typeof PromotionType)[keyof typeof PromotionType];

export const PromotionType = {
  promotion: "promotion",
  demotion: "demotion",
} as const;

export interface Promotion {
  id: string;
  guildId: string;
  staffId: string;
  type: PromotionType;
  fromRankName?: string | null;
  toRankName?: string | null;
  reason?: string | null;
  promotedById: string;
  promotedByUsername: string;
  createdAt: string;
}

export interface Meeting {
  id: string;
  guildId: string;
  title: string;
  description?: string | null;
  hostId: string;
  hostUsername: string;
  scheduledAt: string;
  endedAt?: string | null;
  notes?: string | null;
  isCompleted: boolean;
  attendees: string[];
  createdAt: string;
}

export interface CreateMeetingBody {
  title: string;
  description?: string | null;
  hostId: string;
  hostUsername: string;
  channelId?: string | null;
  scheduledAt: string;
}

export interface GivePremiumBody {
  grantedById: string;
  grantedByUsername: string;
  durationDays?: number | null;
}

export type DiscordAuthCallbackParams = {
  code: string;
  state?: string;
};

export type ListStaffParams = {
  page?: number;
  limit?: number;
  rankId?: string;
  divisionId?: string;
  search?: string;
};

export type ListApplicationsParams = {
  status?: string;
  page?: number;
};

export type ListStrikesParams = {
  staffId?: string;
  active?: boolean;
};

export type ListLoasParams = {
  status?: string;
  staffId?: string;
};

export type ListPromotionsParams = {
  staffId?: string;
  page?: number;
};