# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zenith Staff Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds/{guildId}/strikes:
    get:
      operationId: listStrikes
      tags: [strikes]
      summary: List strikes
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: staffId
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Strikes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strike"
    post:
      operationId: createStrike
      tags: [strikes]
      summary: Issue a strike
      parameters:
        - $ref: "#/components/parameters/GuildId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStrikeBody"
      responses:
        "201":
          description: Strike issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strike"
  /guilds/{guildId}/strikes/{strikeId}/remove:
    post:
      operationId: removeStrike
      tags: [strikes]
      summary: Remove a strike
      parameters:
        - $ref: "#/components/parameters/GuildId"
        - name: strikeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveStrikeBody"
      responses:
        "200":
          description: Strike removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strike"
...
components:
  parameters:
    GuildId:
      name: guildId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Strike:
      type: object
      properties:
        id:
          type: string
        guildId:
          type: string
        staffId:
          type: string
        issuedById:
          type: string
        issuedByUsername:
          type: string
        severity:
          type: string
        reason:
          type: string
        evidence:
          type: string
          nullable: true
        isActive:
          type: boolean
        appealStatus:
          type: string
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required: [id, guildId, staffId, issuedById, issuedByUsername, severity, reason, isActive, appealStatus, issuedAt]
    CreateStrikeBody:
      type: object
      properties:
        staffId:
          type: string
        issuedById:
          type: string
        issuedByUsername:
          type: string
        severity:
          type: string
          enum: [warning, strike, final_warning]
        reason:
          type: string
        evidence:
          type: string
          nullable: true
      required: [staffId, issuedById, issuedByUsername, reason]
    RemoveStrikeBody:
      type: object
      properties:
        removedById:
          type: string
      required: [removedById]
...